+++ src/panparsex/__init__.py
@@ -1,13 +1,14 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2024 Dhruvil Darji
 
-from .core import parse, ParserProtocol, register_parser, get_registry, parse_folder, parse_folder_unified
+from .core import parse, ParserProtocol, register_parser, get_registry, parse_folder, parse_folder_unified, ParsingSummary
 from .types import UnifiedDocument, Section, Chunk, Metadata
 
 __all__ = [
     "parse",
     "parse_folder",
     "parse_folder_unified",
+    "ParsingSummary",
     "ParserProtocol",
     "register_parser",
     "get_registry",